SWARM-1000: Introduced FractionListParser
Motivation
----------
The parsing algorithm used in the FractionList class can be reused in other tooling (the Forge addon for example)

Modifications
-------------
The parsing code on FractionList was extracted to a FractionListParser class, which can be reused without much fuzz.

Result
------
FractionListParser is now available to be used in other tooling solutions. No impact in current code
